+++ playa/content.py
@@ -159,7 +159,7 @@ class GraphicState:
 
     """
 
-    clipping_path: None = None # TODO
+    clipping_path: None = None  # TODO
     linewidth: float = 1
     linecap: int = 0
     linejoin: int = 0
@@ -357,8 +357,10 @@ class ImageObject(ContentObject):
         # region of the page by temporarily altering the CTM.
         return transform_bbox(self.ctm, (0, 0, 1, 1))
 
+
 # Group XObject subtypes. As of PDF 2.0 Transparency is the only defined subtype
 LITERAL_TRANSPARENCY = LIT("Transparency")
+
 
 @dataclass
 class XObjectObject(ContentObject):
@@ -652,7 +654,6 @@ class TextObject(ContentObject):
     _matrix: Union[Matrix, None] = None
     _chars: Union[List[str], None] = None
     _bbox: Union[Rect, None] = None
-    _text_space_bbox: Union[Rect, None] = None
     _next_glyph_offset: Union[Point, None] = None
 
     def __iter__(self) -> Iterator[GlyphObject]:
@@ -714,7 +715,7 @@ class TextObject(ContentObject):
             else:
                 for cid, text in font.decode(obj):
                     glyph_offset = (x, pos) if vert else (pos, y)
-                    disp = font.vdisp(cid) if vert else font.char_width(cid)
+                    disp = font.vdisp(cid) if vert else font.hdisp(cid)
                     disp += scaled_charspace
                     if cid == 32:
                         disp += scaled_wordspace
@@ -742,26 +743,23 @@ class TextObject(ContentObject):
         if self._next_glyph_offset is None:
             self._next_glyph_offset = glyph_offset
 
-    def _calculate_scaled_text_space_bbox(self):
-        if self._text_space_bbox is not None:
-            return self._text_space_bbox
+    def _get_next_glyph_offset(self) -> Point:
+        """Update only the glyph offset without calculating anything else."""
+        if self._next_glyph_offset is not None:
+            return self._next_glyph_offset
         font = self.gstate.font
         fontsize = self.gstate.fontsize
-        rise = self.gstate.rise
         if font is None:
             log.warning(
                 "No font is set, will not update text state or output text: %r TJ",
                 self.args,
             )
-            self._text_space_bbox = BBOX_NONE
             self._next_glyph_offset = self._glyph_offset
-            return self._text_space_bbox
+            return self._next_glyph_offset
         if len(self.args) == 0:
-            self._text_space_bbox = BBOX_NONE
             self._next_glyph_offset = self._glyph_offset
-            return self._text_space_bbox
-        descent = font.get_descent() * fontsize
-        ascent = font.get_ascent() * fontsize
+            return self._next_glyph_offset
+
         horizontal_scaling = self.gstate.scaling * 0.01
         charspace = self.gstate.charspace
         wordspace = self.gstate.wordspace
@@ -770,20 +768,7 @@ class TextObject(ContentObject):
             wordspace = 0
         (x, y) = self._glyph_offset
         pos = y if vert else x
-        if vert:
-            # Because the position vector can be anything for vertical
-            # writing, none of these can be fixed even if we ignore
-            # glyph-specific width, ascent and descent.
-            x0 = x1 = x
-            y0 = y1 = y
-        else:
-            x0 = x1 = x
-            # In horizontal writing by contrast the baseline never
-            # changes, and by convention, even though it's quite
-            # incorrect, we use the descent and rise for the text and
-            # glyph bounding box.  This means y0 and y1 are fixed.
-            y0 = y + descent + rise
-            y1 = y + ascent + rise
+        if not vert:
             # Scale charspace and wordspace, PDF 2.0 section 9.3.2
             charspace *= horizontal_scaling
             wordspace *= horizontal_scaling
@@ -793,38 +778,16 @@ class TextObject(ContentObject):
             else:
                 for cid, _ in font.decode(obj):
                     x, y = (x, pos) if vert else (pos, y)
-                    width = font.char_width(cid)
                     if vert:
                         assert isinstance(font, CIDFont)
-                        adv = font.vdisp(cid) * fontsize
-                        gx0, gy0, gx1, gy1 = font.char_bbox(cid)
-                        gx0 *= fontsize * horizontal_scaling
-                        gx1 *= fontsize * horizontal_scaling
-                        gy0 *= fontsize
-                        gy0 += rise
-                        gy1 *= fontsize
-                        gy1 += rise
-                        x0 = min(x0, x + gx0)
-                        y0 = min(y0, y + gy0)
-                        x1 = max(x1, x + gx1)
-                        y1 = max(y1, y + gy1)
+                        pos += font.vdisp(cid) * fontsize
                     else:
-                        adv = width * fontsize * horizontal_scaling
-                        x1 = x + adv
-                    pos += adv
+                        hdisp = font.hdisp(cid)
+                        pos += hdisp * fontsize * horizontal_scaling
                     pos += charspace
                     if cid == 32:
                         pos += wordspace
-        if self._next_glyph_offset is None:
-            self._next_glyph_offset = (x, pos) if vert else (pos, y)
-        self._text_space_bbox = (x0, y0, x1, y1)
-        return self._text_space_bbox
-
-    def _get_next_glyph_offset(self) -> Point:
-        if self._next_glyph_offset is not None:
-            return self._next_glyph_offset
-        self._calculate_scaled_text_space_bbox()
-        assert self._next_glyph_offset is not None
+        self._next_glyph_offset = (x, pos) if vert else (pos, y)
         return self._next_glyph_offset
 
     @property
@@ -872,7 +835,77 @@ class TextObject(ContentObject):
         if self._bbox is not None:
             return self._bbox
         matrix = mult_matrix(self.line_matrix, self.ctm)
-        self._bbox = transform_bbox(matrix, self._calculate_scaled_text_space_bbox())
+        font = self.gstate.font
+        fontsize = self.gstate.fontsize
+        rise = self.gstate.rise
+        if font is None:
+            log.warning(
+                "No font is set, will not update text state or output text: %r TJ",
+                self.args,
+            )
+            self._bbox = BBOX_NONE
+            self._next_glyph_offset = self._glyph_offset
+            return self._bbox
+        if len(self.args) == 0:
+            self._bbox = BBOX_NONE
+            self._next_glyph_offset = self._glyph_offset
+            return self._bbox
+
+        horizontal_scaling = self.gstate.scaling * 0.01
+        charspace = self.gstate.charspace
+        wordspace = self.gstate.wordspace
+        vert = font.vertical
+        if font.multibyte:
+            wordspace = 0
+        (x, y) = self._glyph_offset
+        pos = y if vert else x
+        x0 = x1 = x
+        y0 = y1 = y + rise
+        fast_path = False
+        if not vert:
+            # Scale charspace and wordspace, PDF 2.0 section 9.3.2
+            charspace *= horizontal_scaling
+            wordspace *= horizontal_scaling
+            # Detect the most frequent case, horizontal writing with
+            # diagonal font.matrix
+            a, b, c, d, e, f = font.matrix
+            if b == 0 and c == 0:
+                fast_path = True
+                y0 += d * font.descent * fontsize
+                y1 += d * font.ascent * fontsize
+        for obj in self.args:
+            if isinstance(obj, (int, float)):
+                pos -= obj * 0.001 * fontsize * horizontal_scaling
+            else:
+                for cid, _ in font.decode(obj):
+                    x, y = (x, pos) if vert else (pos, y)
+                    if vert:
+                        assert isinstance(font, CIDFont)
+                        pos += font.vdisp(cid) * fontsize
+                    else:
+                        hdisp = font.hdisp(cid)
+                        pos += hdisp * fontsize * horizontal_scaling
+                    if fast_path:
+                        x1 = pos
+                    else:
+                        gx0, gy0, gx1, gy1 = font.char_bbox(cid)
+                        gx0 *= fontsize * horizontal_scaling
+                        gx1 *= fontsize * horizontal_scaling
+                        gy0 *= fontsize
+                        gy0 += rise
+                        gy1 *= fontsize
+                        gy1 += rise
+                        x0 = min(x0, x + gx0)
+                        y0 = min(y0, y + gy0)
+                        x1 = max(x1, x + gx1)
+                        y1 = max(y1, y + gy1)
+                    pos += charspace
+                    if cid == 32:
+                        pos += wordspace
+        # Update this because we can!
+        if self._next_glyph_offset is None:
+            self._next_glyph_offset = (x, pos) if vert else (pos, y)
+        self._bbox = transform_bbox(matrix, (x0, y0, x1, y1))
         return self._bbox
 
     @property