+++ playa/font.py
@@ -43,6 +43,7 @@ from playa.parser import (
 )
 from playa.pdftypes import (
     ContentStream,
+    Matrix,
     Rect,
     dict_value,
     int_value,
@@ -56,9 +57,9 @@ from playa.pdftypes import (
 )
 from playa.utils import (
     Point,
-    apply_matrix_norm,
     choplist,
     decode_text,
+    transform_bbox,
 )
 
 log = logging.getLogger(__name__)
@@ -92,7 +93,7 @@ class Font:
         self.descent = num_value(descriptor.get("Descent", -120))
         self.italic_angle = num_value(descriptor.get("ItalicAngle", 0))
         if default_width is None:
-            self.default_width = num_value(descriptor.get("MissingWidth", 0))
+            self.default_width = num_value(descriptor.get("MissingWidth", 1000))
         else:
             self.default_width = default_width
         self.leading = num_value(descriptor.get("Leading", 0))
@@ -100,7 +101,7 @@ class Font:
             self.bbox = rect_value(descriptor["FontBBox"])
         else:
             self.bbox = (0, 0, 0, 0)
-        self.hscale = self.vscale = 0.001
+        self.matrix: Matrix = (0.001, 0, 0, 0.001, 0, 0)
 
         # PDF RM 9.8.1 specifies /Descent should always be a negative number.
         # PScript5.dll seems to produce Descent with a positive number, but
@@ -119,36 +120,11 @@ class Font:
         log.debug("decode with identity: %r", data)
         return ((cid, chr(cid)) for cid in data)
 
-    def get_ascent(self) -> float:
-        """Ascent above the baseline, in text space units"""
-        return self.ascent * self.vscale
-
-    def get_descent(self) -> float:
-        """Descent below the baseline, in text space units; always negative"""
-        return self.descent * self.vscale
-
-    def get_width(self) -> float:
-        """Maximum width of a glyph in this font"""
-        w = self.bbox[2] - self.bbox[0]
-        if w == 0:
-            w = -self.default_width
-        return w * self.hscale
-
-    def get_height(self) -> float:
-        """Maximum height of a glyph in this font"""
-        h = self.bbox[3] - self.bbox[1]
-        if h == 0:
-            h = self.ascent - self.descent
-        return h * self.vscale
-
-    def char_width(self, cid: int) -> float:
-        """Get the width of a character from its CID."""
-        if cid not in self.widths:
-            return self.default_width * self.hscale
-        return self.widths[cid] * self.hscale
-
-    def string_width(self, s: bytes) -> float:
-        return sum(self.char_width(cid) for cid, _ in self.decode(s))
+    def hdisp(self, cid: int) -> float:
+        """Get the horizontal displacement (so-called "width") of a character
+        from its CID."""
+        width = self.widths.get(cid, self.default_width)
+        return self.matrix[0] * width
 
     def vdisp(self, cid: int) -> float:
         """Get vertical displacement for vertical writing mode, in
@@ -181,12 +157,10 @@ class Font:
             the actual bounding box of an individual glyph as
             specified by the font program.
         """
-        descent = self.get_descent()
-        ascent = self.get_ascent()
-        width = self.char_width(cid)
-        x0, y0 = (0, descent)
-        x1, y1 = (width, ascent)
-        return x0, y0, x1, y1
+        width = self.widths.get(cid, self.default_width)
+        # We know the matrix is diagonal
+        a, _, _, d, _, _ = self.matrix
+        return (0, d * self.descent, a * width, d * self.ascent)
 
 
 class SimpleFont(Font):
@@ -318,8 +292,7 @@ class Type1Font(SimpleFont):
             # shall be the font's built-in encoding (see FIXME above)
             return LITERAL_STANDARD_ENCODING
 
-    def char_width(self, cid: int) -> float:
-        """Get the width of a character from its CID."""
+    def _glyph_space_width(self, cid: int) -> float:
         # Commit 6e4f36d <- what's the purpose of this? seems very cursed
         # reverting this would make #76 easy to fix since cid2unicode would only be
         # needed when ToUnicode is absent
@@ -335,12 +308,31 @@ class Type1Font(SimpleFont):
         # Then we can construct `self.widths` directly using `self.encoding`.
         if self.char_widths is not None:
             if cid not in self._cid2unicode:
-                width = self.default_width
-            else:
-                width = self.char_widths.get(self._cid2unicode[cid], self.default_width)
-        else:
-            width = self.widths.get(cid, self.default_width)
-        return width * self.hscale
+                return self.default_width
+            return self.char_widths.get(self._cid2unicode[cid], self.default_width)
+        return self.widths.get(cid, self.default_width)
+
+    def hdisp(self, cid: int) -> float:
+        """Get the horizontal displacement (so-called "width") of a character
+        from its CID."""
+        return self.matrix[0] * self._glyph_space_width(cid)
+
+    def char_bbox(self, cid: int) -> Rect:
+        """Get the standard bounding box for a character from its CID.
+
+        This is, very specifically, `[0 descent width ascent]` in text
+        space units.
+
+        Danger: Not the actual bounding box of the glyph.
+            This is a standardized bounding box for use in text
+            extraction and layout analysis.  It does not correspond to
+            the actual bounding box of an individual glyph as
+            specified by the font program.
+        """
+        width = self._glyph_space_width(cid)
+        # We know the matrix is diagonal
+        a, _, _, d, _, _ = self.matrix
+        return (0, d * self.descent, a * width, d * self.ascent)
 
     def __repr__(self) -> str:
         return "<Type1Font: basefont=%r>" % self.basefont
@@ -390,13 +382,7 @@ class Type3Font(SimpleFont):
             self.fontname = decode_text(fontname)
         else:
             self.fontname = "unknown"
-        # However in many cases the name will be clearly "subset-like"
-        # so set basefont to the "base font" part of it
-        _, _, basefont = self.fontname.partition("+")
-        if basefont:
-            self.basefont = basefont
-        else:
-            self.basefont = self.fontname
+        self.basefont = self.fontname
 
         if "FontMatrix" in spec:  # it is actually required though
             self.matrix = matrix_value(spec["FontMatrix"])
@@ -406,11 +392,9 @@ class Type3Font(SimpleFont):
         if "FontBBox" in spec:  # it is also required though
             self.bbox = rect_value(spec["FontBBox"])
             # otherwise it was set in SimpleFont.__init__
-        # set ascent/descent from the bbox (they *could* be in the
+        # Set ascent/descent from the bbox (they *could* be in the
         # descriptor but this is very unlikely)
         _, self.descent, _, self.ascent = self.bbox
-        # determine the actual height/width applying transformation
-        (self.hscale, self.vscale) = apply_matrix_norm(self.matrix, (1, 1))
 
     def get_implicit_encoding(
         self, descriptor: Dict[str, PDFObject]
@@ -419,6 +403,20 @@ class Type3Font(SimpleFont):
         # codes to glyph names shall be entirely defined by its
         # Encoding entry, which is required in this case.
         return {}
+
+    def char_bbox(self, cid: int) -> Rect:
+        """Get the standard bounding box for a character from its CID.
+
+        This is the smallest rectangle enclosing [0 descent width
+        ascent] after the font matrix has been applied.
+
+        Danger: Not the actual bounding box of the glyph (but almost).
+            The descent and ascent here are from the **font** and not
+            from the individual **glyph** so this will be somewhat
+            larger than the actual bounding box.
+        """
+        width = self.widths.get(cid, self.default_width)
+        return transform_bbox(self.matrix, (0, self.descent, width, self.ascent))
 
     def __repr__(self) -> str:
         return "<Type3Font>"
@@ -665,7 +663,7 @@ class CIDFont(Font):
 
         Returns 0 for horizontal writing, for obvious reasons.
         """
-        return self.vdisps.get(cid, self.default_vdisp) * self.vscale
+        return self.matrix[3] * self.vdisps.get(cid, self.default_vdisp)
 
     def position(self, cid: int) -> Tuple[float, float]:
         """Get position vector for vertical writing mode, in text
@@ -689,7 +687,9 @@ class CIDFont(Font):
 
         """
         vx, vy = self.positions.get(cid, self.default_position)
-        return vx * self.hscale, vy * self.vscale
+        # We know that the matrix is diagonal here
+        a, _, _, d, _, _ = self.matrix
+        return a * vx, d * vy
 
     def char_bbox(self, cid: int) -> Rect:
         """Get the standard bounding box for a character from its CID.
@@ -704,22 +704,21 @@ class CIDFont(Font):
             specified by the font program.
 
         """
-        descent = self.get_descent()
-        ascent = self.get_ascent()
-        width = self.char_width(cid)
+        width = self.widths.get(cid, self.default_width)
+        # We know that the matrix is diagonal here
+        a, _, _, d, _, _ = self.matrix
         if self.vertical:
-            vx, vy = self.position(cid)
+            vx, vy = self.positions.get(cid, self.default_position)
             # Horizontal offset for glyph origin vs. text
             # space origin.
             vx = -vx
             # Vertical offset for glyph origin
             vy = -vy
             # Find glyph bbox
-            x0 = vx
-            y0 = vy + descent
-            x1 = vx + width
-            y1 = vy + ascent
-        else:
-            x0, y0 = (0, descent)
-            x1, y1 = (width, ascent)
-        return x0, y0, x1, y1
+            return (
+                a * vx,
+                d * (vy + self.descent),
+                a * (vx + width),
+                d * (vy + self.ascent),
+            )
+        return (0, d * self.descent, a * width, d * self.ascent)